+++ test/UtilsTest.js
@@ -1,10 +1,10 @@
-/* global expect */
+/* global expect,describe,it */
 /* eslint no-console: 0 */
 /* eslint-env node, mocha */
 
 'use strict';
 
-import {uuid4} from 'Utils';
+import {uuid4} from '../src/utils';
 
 describe('Utils', () => {